SCRUM-22 Created a file for the comparison story, also connects to the story on creating graphs for mean and std curves (SCRUM-37).
It gets the latest runs from ppo folder and baseline folder, then gets the rewards from the monitor files.
With that info, i calculate all the metrics for comparison and also got a bunch of functions that create different graphs.
Saves a small summary to the md file in comparison dir